# apps/core/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoreX Service API",
    "description": "HTTP 状态码始终为 200；业务结果见响应体 body.code（200000=成功）。错误码规则见 /guide/error-codes.md。Exception.details 字段仅在开发环境返回。鉴权：JWT Bearer；Apifox 环境变量统一使用 {{token}}（来自 signin.data.token）。密码登录 / 注册 / 发 OTP 前须先 POST /api/v1/auth/captcha 获取 captchaKey（go-captcha 行为验证码）。",
    "contact": {
      "name": "CoreX Team",
      "email": "15638470820@163.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "development"
    },
    {
      "url": "https://api.example.com",
      "description": "production"
    }
  ],
  "paths": {
...
    "/api/v1/upload/finalize": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "完成上传",
        "description": "校验 chunk 表分片齐全并流式计算整文件 SHA-256；不合并落盘。返回按 asset id 的下载 URL。需要 JWT。",
        "operationId": "upload.finalize",
        "requestBody": {
          "description": "上传会话 id",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeP"
              },
              "example": {
                "id": "550e8400-e29b-41d4-a716-446655440000"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "未登录或校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/upload/prepare": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "初始化上传",
        "description": "创建上传会话。`hash` 可选。文件秒传：本人已完成直接返回；他人已完成则为当前用户克隆 COMPLETED（共享 CAS，克隆行默认 PRIVATE）。\n\n**可见性**（`visibility`，默认 `PRIVATE`）：\n- `PRIVATE`：仅创建者可按 id 下载（需 JWT）\n- `PUBLIC`：匿名或已登录均可按 id 下载（`GET /asset/{id}` 对 Auth 放行）\n- `RESTRICTED`：创建者 + `viewers`（用户 UUID 列表，最多 100）可下载（需 JWT）\n\n上传会话写操作（chunk/hash/finalize 等）仍仅创建者。需要 JWT。",
        "operationId": "upload.prepare",
        "requestBody": {
          "description": "初始化参数；hash / visibility 可省略（默认 PRIVATE）",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareP"
              },
              "example": {
                "chunk": 1048576,
                "mime": "application/pdf",
                "name": "sample.pdf",
                "size": 1048576,
                "visibility": "PRIVATE"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "未登录或参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/upload/progress/{id}": {
      "get": {
        "tags": [
          "Upload"
        ],
        "summary": "查询上传进度",
        "description": "SUPERSEDED 时 progress=100 并返回 superseded（目标 COMPLETED）。需要 JWT。",
        "operationId": "upload.progress",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "上传会话 ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "未登录或会话不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exception"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Exception": {
        "type": "object",
        "description": "异常响应结构",
        "required": [
          "code",
          "success",
          "msg",
          "timestamp"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "业务错误码 (非0表示各种业务错误)"
          },
          "data": {
            "type": [
              "object",
              "null"
            ]
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "description": "详细错误信息（仅非生产环境返回）"
          },
          "msg": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "FinalizeP": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        },
        "example": {
          "id": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
...
      "PrepareP": {
        "type": "object",
        "required": [
          "name",
          "size",
          "mime",
          "chunk"
        ],
        "properties": {
          "chunk": {
            "type": "integer",
            "format": "int32",
            "description": "分片大小（字节），须在 10MB~100MB",
            "example": 1048576,
            "minimum": 0
          },
          "hash": {
            "type": [
              "string",
              "null"
            ],
            "description": "整文件 SHA-256；可省略，稍后再 `PATCH /upload/hash` 绑定",
            "example": "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
          },
          "index": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "租户内列表排序；省略则默认 0",
            "example": 0
          },
          "mime": {
            "type": "string",
            "example": "application/pdf"
          },
          "name": {
            "type": "string",
            "example": "sample.pdf"
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "example": 1048576,
            "minimum": 0
          },
          "tenantID": {
            "type": [
              "string",
              "null"
            ],
            "description": "租户 ID（租户模型未接线前可选）",
            "example": "tenant-demo"
          },
          "viewers": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "RESTRICTED 时的可访问用户 id 列表；其它可见性忽略",
            "example": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          },
          "visibility": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Visibility",
                "description": "可见性；省略则 PRIVATE"
              }
            ]
          }
        },
        "example": {
          "chunk": 1048576,
          "mime": "application/pdf",
          "name": "sample.pdf",
          "size": 1048576
        }
      },
...
      "Visibility": {
        "type": "string",
        "description": "资产可见性：私有 / 公开 / 指定用户",
        "enum": [
          "PRIVATE",
          "PUBLIC",
          "RESTRICTED"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录 `POST /api/v1/auth/signin` 返回的 `data.token`。\n\n请求头：`Authorization: Bearer <token>`。\n\n**Apifox**：鉴权组件 Token 请填 `{{token}}`（勿用 `{{bearerToken}}`）；环境变量名统一为 `token`，可在登录接口后置操作写入。",
        "x-default": "{{token}}",
        "x-apifox-default": "{{token}}"
      }
    }
  },
...
}